// server/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  ADMIN: 'ADMIN',
  USER: 'USER'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const UserStatus = {
  PENDING: 'PENDING',
  ACTIVE: 'ACTIVE',
  SUSPENDED: 'SUSPENDED',
  DELETED: 'DELETED'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const CarCategory = {
  Sedan: 'Sedan',
  SUV: 'SUV',
  Hatchback: 'Hatchback',
  Coupe: 'Coupe',
  Convertible: 'Convertible',
  Sports: 'Sports',
  Luxury: 'Luxury',
  Supercar: 'Supercar',
  MUV: 'MUV',
  MPV: 'MPV',
  Van: 'Van',
  PickupTruck: 'PickupTruck',
  Electric: 'Electric',
  Hybrid: 'Hybrid',
  OffRoad: 'OffRoad'
} as const

export type CarCategory = (typeof CarCategory)[keyof typeof CarCategory]


export const FuelType = {
  Petrol: 'Petrol',
  Diesel: 'Diesel',
  Electric: 'Electric',
  Hybrid: 'Hybrid',
  CNG: 'CNG',
  LPG: 'LPG',
  Hydrogen: 'Hydrogen'
} as const

export type FuelType = (typeof FuelType)[keyof typeof FuelType]


export const Transmission = {
  Manual: 'Manual',
  Automatic: 'Automatic',
  SemiAutomatic: 'SemiAutomatic',
  AMT: 'AMT',
  CVT: 'CVT',
  DCT: 'DCT'
} as const

export type Transmission = (typeof Transmission)[keyof typeof Transmission]


export const Location = {
  Albuquerque: 'Albuquerque',
  Atlanta: 'Atlanta',
  Austin: 'Austin',
  Baltimore: 'Baltimore',
  Boston: 'Boston',
  Charlotte: 'Charlotte',
  Chicago: 'Chicago',
  Cleveland: 'Cleveland',
  Colorado_Springs: 'Colorado_Springs',
  Columbus: 'Columbus',
  Dallas: 'Dallas',
  Denver: 'Denver',
  Detroit: 'Detroit',
  El_Paso: 'El_Paso',
  Fort_Worth: 'Fort_Worth',
  Fresno: 'Fresno',
  Houston: 'Houston',
  Indianapolis: 'Indianapolis',
  Jacksonville: 'Jacksonville',
  Kansas_City: 'Kansas_City',
  Las_Vegas: 'Las_Vegas',
  Long_Beach: 'Long_Beach',
  Los_Angeles: 'Los_Angeles',
  Memphis: 'Memphis',
  Miami: 'Miami',
  Milwaukee: 'Milwaukee',
  Minneapolis: 'Minneapolis',
  Nashville: 'Nashville',
  New_Orleans: 'New_Orleans',
  New_York: 'New_York',
  Oakland: 'Oakland',
  Oklahoma_City: 'Oklahoma_City',
  Omaha: 'Omaha',
  Orlando: 'Orlando',
  Philadelphia: 'Philadelphia',
  Phoenix: 'Phoenix',
  Pittsburgh: 'Pittsburgh',
  Portland: 'Portland',
  Sacramento: 'Sacramento',
  San_Antonio: 'San_Antonio',
  San_Diego: 'San_Diego',
  San_Francisco: 'San_Francisco',
  San_Jose: 'San_Jose',
  Seattle: 'Seattle',
  Tampa: 'Tampa',
  Tucson: 'Tucson',
  Tulsa: 'Tulsa',
  Virginia_Beach: 'Virginia_Beach',
  Washington_DC: 'Washington_DC',
  Wichita: 'Wichita'
} as const

export type Location = (typeof Location)[keyof typeof Location]


export const BookingStatus = {
  PENDING: 'PENDING',
  CONFIRMED: 'CONFIRMED',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED'
} as const

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus]


export const PaymentMethod = {
  CREDIT_CARD: 'CREDIT_CARD',
  DEBIT_CARD: 'DEBIT_CARD',
  PAYPAL: 'PAYPAL',
  CASH: 'CASH',
  UPI: 'UPI'
} as const

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod]